third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-percentage-expected.txt

This is a testharness.js-based test.
PASS Animating a custom property of type <length-percentage>
PASS Animating a custom property of type <length-percentage> with a single keyframe
PASS Animating a custom property of type <length-percentage> with additivity
PASS Animating a custom property of type <length-percentage> with a single keyframe and additivity
FAIL Animating a custom property of type <length-percentage> with iterationComposite assert_equals: expected "calc(250% + 25px)" but got "calc(50% + 25px)"
Harness: the test ran to completion.
